multibot: Split the GAME_DROIDINFO wire format into multibot_serde.h
